enable connecting wallet to user via oauth writeOnce
