+++ packages/web/src/pages/oauth-login-page/hooks.ts
@@ -1,10 +1,36 @@
-import { useMemo } from 'react'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 
+import {
+  accountSelectors,
+  CommonState,
+  encodeHashId,
+  ErrorLevel,
+  Name,
+  statusIsNotFinalized,
+  User
+} from '@audius/common'
 import * as queryString from 'query-string'
-import { useLocation } from 'react-router-dom'
+import { useDispatch, useSelector } from 'react-redux'
+import { useHistory, useLocation } from 'react-router-dom'
+
+import { make, useRecord } from 'common/store/analytics/actions'
+import { audiusBackendInstance } from 'services/audius-backend/audius-backend-instance'
+import * as errorActions from 'store/errors/actions'
+import { reportToSentry } from 'store/errors/reportToSentry'
 
 import { messages } from './messages'
-import { getIsRedirectValid, isValidApiKey } from './utils'
+import {
+  authWrite,
+  connectUserToDashboardWallet,
+  formOAuthResponse,
+  getDeveloperApp,
+  getIsAppAuthorized,
+  getIsRedirectValid,
+  isValidApiKey,
+  validateWriteOnceParams,
+  WriteOnceParams
+} from './utils'
+const { getAccountUser, getAccountStatus } = accountSelectors
 
 export const useParsedQueryParams = () => {
   const { search } = useLocation()
@@ -16,7 +42,9 @@ export const useParsedQueryParams = () => {
     app_name: appName,
     response_mode: responseMode,
     api_key: apiKey,
-    origin
+    origin,
+    tx,
+    ...rest
   } = queryString.parse(search)
 
   const parsedRedirectUri = useMemo<'postmessage' | URL | null>(() => {
@@ -48,33 +76,55 @@ export const useParsedQueryParams = () => {
     return null
   }, [origin])
 
-  let initError: string | null = null
-  if (isRedirectValid === false) {
-    initError = messages.redirectURIInvalidError
-  } else if (parsedRedirectUri === 'postmessage' && !parsedOrigin) {
-    // Only applicable if redirect URI set to `postMessage`
-    initError = messages.originInvalidError
-  } else if (scope !== 'read' && scope !== 'write') {
-    initError = messages.scopeError
-  } else if (
-    responseMode &&
-    responseMode !== 'query' &&
-    responseMode !== 'fragment'
-  ) {
-    initError = messages.responseModeError
-  } else if (scope === 'read') {
-    // Read scope-specific validations:
-    if (!appName && !apiKey) {
-      initError = messages.missingAppNameError
+  const { error, txParams } = useMemo(() => {
+    let error: string | null = null
+    let txParams: WriteOnceParams | null = null // Only used for scope=write_once
+    if (isRedirectValid === false) {
+      error = messages.redirectURIInvalidError
+    } else if (parsedRedirectUri === 'postmessage' && !parsedOrigin) {
+      // Only applicable if redirect URI set to `postMessage`
+      error = messages.originInvalidError
+    } else if (
+      scope !== 'read' &&
+      scope !== 'write' &&
+      scope !== 'write_once'
+    ) {
+      error = messages.scopeError
+    } else if (
+      responseMode &&
+      responseMode !== 'query' &&
+      responseMode !== 'fragment'
+    ) {
+      error = messages.responseModeError
+    } else if (scope === 'read') {
+      // Read scope-specific validations:
+      if (!appName && !apiKey) {
+        error = messages.missingAppNameError
+      }
+    } else if (scope === 'write') {
+      // Write scope-specific validations:
+      if (!apiKey) {
+        error = messages.missingApiKeyError
+      } else if (!isValidApiKey(apiKey)) {
+        error = messages.invalidApiKeyError
+      }
+    } else if (scope === 'write_once') {
+      // Write-once scope-specific validations:
+      const { error: writeOnceParamsError, txParams: txParamsRes } =
+        validateWriteOnceParams({
+          tx,
+          params: rest
+        })
+      txParams = txParamsRes
+
+      if (writeOnceParamsError) {
+        error = writeOnceParamsError
+      }
     }
-  } else if (scope === 'write') {
-    // Write scope-specific validations:
-    if (!apiKey) {
-      initError = messages.missingApiKeyError
-    } else if (!isValidApiKey(apiKey)) {
-      initError = messages.invalidApiKeyError
-    }
-  }
+    return { txParams, error }
+    // This is exhaustive despite what eslint thinks:
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isRedirectValid, parsedOrigin, parsedRedirectUri, search])
 
   return {
     apiKey,
@@ -87,6 +137,334 @@ export const useParsedQueryParams = () => {
     parsedRedirectUri,
     isRedirectValid,
     parsedOrigin,
+    error,
+    tx,
+    txParams
+  }
+}
+
+export const useOAuthSetup = () => {
+  const record = useRecord()
+  const history = useHistory()
+  const dispatch = useDispatch()
+  const {
+    appName: queryParamAppName,
+    apiKey,
+    scope,
+    redirectUri,
+    state,
+    responseMode,
+    origin: originParam,
+    parsedRedirectUri,
+    isRedirectValid,
+    parsedOrigin,
+    txParams,
+    tx,
     error: initError
+  } = useParsedQueryParams()
+  const accountIsLoading = useSelector((state: CommonState) => {
+    const status = getAccountStatus(state)
+    return statusIsNotFinalized(status)
+  })
+  const account = useSelector(getAccountUser)
+  const isLoggedIn = Boolean(account)
+  const [userEmail, setUserEmail] = useState<string | null>(null)
+  const [queryParamsError, setQueryParamsError] = useState<string | null>(
+    initError
+  )
+  /** The fetched developer app name if write OAuth (we use `queryParamAppName` if read or writeOnce OAuth and no API key is given) */
+  const [registeredDeveloperAppName, setRegisteredDeveloperAppName] =
+    useState<string>()
+  const appName = registeredDeveloperAppName ?? queryParamAppName
+
+  const [userAlreadyWriteAuthorized, setUserAlreadyWriteAuthorized] =
+    useState<boolean>()
+
+  const loading =
+    accountIsLoading ||
+    (apiKey &&
+      (registeredDeveloperAppName === undefined ||
+        userAlreadyWriteAuthorized === undefined))
+
+  useEffect(() => {
+    if (!queryParamsError) {
+      record(
+        make(Name.AUDIUS_OAUTH_START, {
+          redirectUriParam:
+            parsedRedirectUri === 'postmessage' ? 'postmessage' : redirectUri!,
+          originParam,
+          responseMode,
+          scope: scope!,
+          apiKeyParam: apiKey,
+          appId: (apiKey || appName)!
+        })
+      )
+    }
+  }, [
+    appName,
+    originParam,
+    parsedRedirectUri,
+    queryParamsError,
+    record,
+    redirectUri,
+    responseMode,
+    apiKey,
+    scope
+  ])
+
+  useEffect(() => {
+    const fetchDeveloperAppName = async () => {
+      if (!apiKey || queryParamsError) {
+        return
+      }
+      let developerApp
+      try {
+        developerApp = await getDeveloperApp(apiKey as string)
+      } catch {
+        setQueryParamsError(messages.invalidApiKeyError)
+        return
+      }
+      if (!developerApp) {
+        setQueryParamsError(messages.invalidApiKeyError)
+        return
+      }
+      setRegisteredDeveloperAppName(developerApp.name)
+    }
+    fetchDeveloperAppName()
+  }, [apiKey, queryParamAppName, queryParamsError, scope])
+
+  useEffect(() => {
+    const getAndSetEmail = async () => {
+      let email: string
+      try {
+        email = await audiusBackendInstance.getUserEmail()
+      } catch {
+        setUserEmail(null)
+        dispatch(
+          errorActions.handleError({
+            name: 'Get User Email',
+            message: 'Get User Email',
+            shouldRedirect: true
+          })
+        )
+        return
+      }
+      setUserEmail(email)
+    }
+    if (isLoggedIn) {
+      getAndSetEmail()
+    } else {
+      setUserEmail(null)
+    }
+  }, [history, isLoggedIn, dispatch])
+
+  const formResponseAndRedirect = useCallback(
+    async ({
+      account,
+      grantCreated,
+      onError
+    }: {
+      account: User
+      grantCreated?: boolean | undefined
+      onError: ({
+        isUserError,
+        errorMessage,
+        error
+      }: {
+        isUserError: boolean
+        errorMessage: string
+        error?: Error
+      }) => void
+    }) => {
+      const jwt = await formOAuthResponse({
+        account,
+        userEmail,
+        onError: () => {
+          dispatch(
+            errorActions.handleError({
+              name: 'Form OAuth Response Error',
+              message: 'Form OAuth Response Error',
+              shouldRedirect: true
+            })
+          )
+        }
+      })
+      if (jwt == null) {
+        onError({ isUserError: false, errorMessage: messages.miscError })
+        return
+      }
+      if (isRedirectValid === true) {
+        record(
+          make(Name.AUDIUS_OAUTH_COMPLETE, {
+            appId: (apiKey || appName)!,
+            scope: scope!,
+            alreadyAuthorized: grantCreated
+          })
+        )
+        if (parsedRedirectUri === 'postmessage') {
+          if (parsedOrigin) {
+            if (!window.opener) {
+              onError({
+                isUserError: false,
+                errorMessage: messages.noWindowError
+              })
+            } else {
+              window.opener.postMessage(
+                { state, token: jwt },
+                parsedOrigin.origin
+              )
+            }
+          }
+        } else {
+          if (responseMode && responseMode === 'query') {
+            if (state != null) {
+              parsedRedirectUri!.searchParams.append('state', state as string)
+            }
+            parsedRedirectUri!.searchParams.append('token', jwt)
+          } else {
+            const statePart = state != null ? `state=${state}&` : ''
+            parsedRedirectUri!.hash = `#${statePart}token=${jwt}`
+          }
+          window.location.href = parsedRedirectUri!.toString()
+        }
+      }
+    },
+    [
+      isRedirectValid,
+      parsedOrigin,
+      parsedRedirectUri,
+      record,
+      responseMode,
+      state,
+      userEmail,
+      apiKey,
+      appName,
+      scope,
+      dispatch
+    ]
+  )
+
+  useEffect(() => {
+    const getInitialAuthorizationStatus = async () => {
+      if (queryParamsError || !apiKey || !isLoggedIn) {
+        setUserAlreadyWriteAuthorized(false)
+        return
+      }
+      let appAlreadyAuthorized
+      try {
+        appAlreadyAuthorized = await getIsAppAuthorized({
+          userId: encodeHashId(account!.user_id), // We know account exists because isLoggedIn is true
+          apiKey: apiKey as string
+        })
+      } catch (e) {
+        if (e instanceof Error) {
+          reportToSentry({ level: ErrorLevel.Error, error: e })
+        }
+        dispatch(
+          errorActions.handleError({
+            name: 'Get Is App Authorized',
+            message: 'Get Is App Authorized',
+            shouldRedirect: true
+          })
+        )
+        return
+      }
+      setUserAlreadyWriteAuthorized(appAlreadyAuthorized)
+    }
+    getInitialAuthorizationStatus()
+  }, [
+    account,
+    apiKey,
+    formResponseAndRedirect,
+    history,
+    isLoggedIn,
+    queryParamsError,
+    scope,
+    dispatch
+  ])
+
+  const authorize = async ({
+    account,
+    onError
+  }: {
+    account: User
+    onError: ({
+      isUserError,
+      errorMessage,
+      error
+    }: {
+      isUserError: boolean
+      errorMessage: string
+      error?: Error
+    }) => void
+  }) => {
+    let shouldCreateWriteGrant = false
+
+    if (scope === 'write') {
+      try {
+        shouldCreateWriteGrant = !(await getIsAppAuthorized({
+          userId: encodeHashId(account.user_id),
+          apiKey: apiKey as string
+        }))
+        if (shouldCreateWriteGrant) {
+          await authWrite({
+            userId: encodeHashId(account.user_id),
+            appApiKey: apiKey as string
+          })
+        }
+      } catch (e: unknown) {
+        let error = 'Creating write grant failed'
+        if (typeof e === 'string') {
+          error = e.toUpperCase()
+        } else if (e instanceof Error) {
+          error = e.message
+        }
+        onError({
+          isUserError: false,
+          errorMessage: messages.miscError,
+          error: e instanceof Error ? e : new Error(error)
+        })
+        return
+      }
+    } else if (scope === 'write_once') {
+      // Note: Tx = 'connect_dashboard_wallet' since that's the only option available right now for write_once scope
+      try {
+        await connectUserToDashboardWallet({
+          userId: encodeHashId(account.user_id),
+          wallet: txParams!.wallet,
+          walletSignature: txParams!.wallet_signature
+        })
+      } catch (e: unknown) {
+        let error = 'Connecting dashboard wallet failed'
+        if (typeof e === 'string') {
+          error = e
+        } else if (e instanceof Error) {
+          error = e.message
+        }
+        onError({
+          isUserError: false,
+          errorMessage: messages.miscError,
+          error: e instanceof Error ? e : new Error(error)
+        })
+        return
+      }
+    }
+    await formResponseAndRedirect({
+      account,
+      grantCreated: shouldCreateWriteGrant,
+      onError
+    })
+  }
+
+  return {
+    scope,
+    queryParamsError,
+    loading,
+    userAlreadyWriteAuthorized,
+    apiKey,
+    appName,
+    userEmail,
+    authorize,
+    tx
   }
 }
