+++ packages/web/src/pages/oauth-login-page/messages.ts
@@ -24,6 +24,9 @@ export const messages = {
     'Whoops, this is an invalid link (redirect URI missing or invalid).',
   missingAppNameError: 'Whoops, this is an invalid link (app name missing).',
   scopeError: `Whoops, this is an invalid link (scope missing or invalid).`,
+  writeOnceParamsError:
+    'Whoops, this is an invalid link (transaction params missing or invalid).',
+  writeOnceTxError: `Whoops, this is an invalid link ('tx' missing or invalid).`,
   missingFieldError: 'Whoops, you must enter both your email and password.',
   originInvalidError:
     'Whoops, this is an invalid link (redirect URI is set to `postMessage` but origin is missing).',
